Require Magenta GPU package for now
We need to resolve the issue of GPU dependencies with our
packages. The likely step will be to drop the package requirement (as
we can't know what the user wants at packaging time) and move that
into the model defs, where we can at least provide some help to users,
either via a message or by running subsequent pip installs based on
the version of TensorFlow installed).